Use type aliases to unify API definitions
This commit annotates the return types via type aliases to ensure that the
low-level ffi definitions correclty represent the intent in terms of input output.

Given that we also have definitions that return "pure" number types, like the
length variants, this approach makes it clear the meaning of each return type.