Add a disassembler test for the trampoline
This commit adds a disassembler test for the trampoline allowing which
should make it easier to inspect code generation related changes.

When the trampoline-generating code is changed, this code will assert
that the generated code is equal to the `trampoline.wat` file on disk,
failing if they are different.

To update the `trampoline.wat` file, one can run the integration tests
as

    TRAMPOLINE_TEST_BLESS=1 cargo test