Ensure that `consumer.wat` always uses the entire API surface
In the process of tightening the snapshot tests, I realized that if a
new API is added to the provider, that won't get caught by the
snapshot tests. This commit streamlines the handling of the API
function imports asserting tha the `consumer.wat` always uses all the
functions that are transformed/renamed by the trampoline.